backend/pyproject.toml
[tool.poetry]
name = "pr-review-app"
version = "1.0.0"
description = "PR Review Application supporting GitHub and GitLab with local LLM analysis"
authors = ["Your Name <you@example.com>"]
readme = "../README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.118.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
sqlalchemy = "2.0.25"
    python-gitlab = "^3.15.0"
PyGithub = "^1.59.0"
httpx = "0.26.0"
pydantic = "^2.8.2"
pydantic-settings = "2.1.0"
python-dotenv = "1.0.0"
litellm = "1.77.7"
dynaconf = "3.2.4"
loguru = "0.7.2"
Jinja2 = "3.1.6"
tenacity = "8.2.3"
tiktoken = "0.8.0"
PyYAML = "6.0.1"
aiohttp = "3.12.15"
GitPython = "3.1.41"
retry = "0.9.2"
ujson = "5.8.0"
html2text = "2024.2.26"
anthropic = ">=0.69.0"
openai = ">=1.55.3"
starlette-context = "0.3.6"
pystray = "^0.19.5"
pillow = "^12.1.0"
pr-agent = {git = "https://github.com/Codium-ai/pr-agent.git"}

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
